src/python/dev-requirements.txt
# Required dependencies
.

# Dev dependencies
selenium>=3.141.0
pytest-cov>=2.6.1
flake8>=3.7.7
jupyter>=1.0.0
ipywidgets>=7.4.2